Add files via upload
Netflix is an American subscription video on-demand over-the-top streaming service. The service primarily distributes original and acquired films and television shows from various genres, and it is available internationally in multiple languages (ir.netflix.net). Utilizing Power BI, an effective business intelligence tool, I extensively explore Netflix data to reveal valuable insights regarding the highest-rated shows and movies, user inclinations, and demographic patterns among the audience. This examination serves as a valuable resource for content creators, distributors, and marketers, empowering them to make well-informed decisions regarding content production, marketing strategies, and distribution channels. The use of interactive visualizations and data-driven analysis facilitates a comprehensive comprehension of the preferences and actions of the vast user base that depends on Netflix on a daily basis.